refactor: 최근 본 상품 컴포넌트 분리
