+++ src/components/RecentlyClicked/index.js
@@ -1,0 +1,81 @@
+import { useState } from "react";
+import styled from "styled-components";
+import { primaryFont } from "styles/common";
+import { GoBookmark } from "react-icons/go";
+import ImageSlide from "./image-slide";
+import ScrollToTop from "./scroll-to-top";
+
+const RecentlyClicked = () => {
+	// 추후 API로 데이터 들어오면 수정
+	const [likes, setLikes] = useState(0);
+
+	return (
+		<S.Container>
+			<S.Top>
+				찜한 상품 <GoBookmark color="white" size="22" /> {likes}
+			</S.Top>
+			<S.Middle>
+				<div className="title">최근 본 상품</div>
+				<ImageSlide />
+			</S.Middle>
+			<ScrollToTop />
+		</S.Container>
+	);
+};
+
+export default RecentlyClicked;
+
+const Container = styled.div`
+	z-index: 100;
+	width: 152.5px;
+	height: 345px;
+	border: 1px solid;
+	border-color: ${({ theme }) => theme.PALETTE.primary};
+	position: fixed;
+	top: 250px;
+	left: 90%;
+	${primaryFont}
+	text-align: center;
+	font-size: 18px;
+	background-color: ${({ theme }) => theme.PALETTE.white};
+`;
+
+const Top = styled.div`
+	display: inline-block;
+	vertical-align: middle;
+	padding: 0 15px;
+	height: 52px;
+	background-color: ${({ theme }) => theme.PALETTE.primary};
+	color: ${({ theme }) => theme.PALETTE.white};
+	font-weight: 400;
+	line-height: 52px;
+	.heart {
+		font-size: 17px;
+		font-weight: 700;
+	}
+	border-bottom: 3px solid;
+
+	svg {
+		vertical-align: middle;
+	}
+`;
+
+const Middle = styled.div`
+	height: 242px;
+	padding: 20px 0;
+	.title {
+		margin-bottom: 10px;
+	}
+	button {
+		border: none;
+		background-color: transparent;
+		font-size: 25px;
+		display: block;
+	}
+`;
+
+const S = {
+	Container,
+	Top,
+	Middle,
+};
