+++ src/components/RecentlyClicked/scroll-to-top.js
@@ -1,0 +1,30 @@
+import { useState } from "react";
+import styled from "styled-components";
+
+const ScrollToTop = () => {
+	// Scroll to Top
+	const [scrollY, setScrollY] = useState(0);
+	// console.log("Y", scrollY);
+
+	const handleScrollToTop = () => {
+		window.scrollTo({ top: 0, behavior: "smooth" });
+		setScrollY(0);
+	};
+	return (
+		<>
+			<S.ScollBtn onClick={handleScrollToTop}>TOP</S.ScollBtn>
+		</>
+	);
+};
+
+export default ScrollToTop;
+
+const ScollBtn = styled.div`
+	height: 45px;
+	border-top: 1px solid;
+	border-color: ${({ theme }) => theme.PALETTE.primary};
+	line-height: 45px;
+	cursor: pointer;
+`;
+
+const S = { ScollBtn };
